Implement ArrowsLinearRing
