+++ osmdroid-android/src/main/java/org/osmdroid/views/overlay/LinearRing.java
@@ -45,21 +45,19 @@ class LinearRing implements SegmentClipper.SegmentClippable{
 	private final ArrayList<PointL> mProjectedPoints = new ArrayList<>();
 	private final PointL mLatestPathPoint = new PointL();
 	private SegmentClipper mSegmentClipper;
+	private SegmentClipper.SegmentClippable mSegmentClipperChild;
 	private final Path mPath;
 	private boolean mIsNextAMove;
 	private boolean mPrecomputed;
 	private boolean isHoritonalRepeating = true;
 	private boolean isVerticalRepeating  = true;
 
-	private final ArrayList<Path> mDirectionalArrows = new ArrayList<>();
-	private static final float DEFAULT_ARROW_LENGTH = 15f;
-	private static final boolean DEFAULT_INVERT_ARROWS = false;
-	private boolean mDrawDirectionalArrows = false;
-	private boolean mInvertDirectionalArrows = DEFAULT_INVERT_ARROWS;
-	private float mDirectionalArrowLength = DEFAULT_ARROW_LENGTH;
-
 	public LinearRing(final Path pPath) {
 		mPath = pPath;
+	}
+
+	void setSegmentClipperChild(final SegmentClipper.SegmentClippable segmentClipperChild) {
+		mSegmentClipperChild = segmentClipperChild;
 	}
 
 	@Override
@@ -81,9 +79,10 @@ class LinearRing implements SegmentClipper.SegmentClippable{
 		}
 	}
 
+	boolean getIsNextMove() { return mIsNextAMove; }
+
 	void clearPath() {
 		mOriginalPoints.clear();
-		mDirectionalArrows.clear();
 		mPrecomputed = false;
 	}
 
@@ -92,42 +91,17 @@ class LinearRing implements SegmentClipper.SegmentClippable{
 		mPrecomputed = false;
 	}
 
+	PointL getLatestPathPoint() { return mLatestPathPoint; }
+
 	ArrayList<GeoPoint> getPoints(){
 		return mOriginalPoints;
 	}
 
+	ArrayList<PointL> getProjectedPoints() { return mProjectedPoints; }
+
 	void setPoints(final List<GeoPoint> points) {
 		clearPath();
 		mOriginalPoints.addAll(points);
-	}
-
-	/**
-	 * A directional arrow is a single arrow drawn in the middle of two points of a polyline to
-	 * provide a visual cue for direction of movement between the two points.
-	 *
-	 * By default the arrows always point towards the lower index as the list of GeoPoints are
-	 * processed. The direction the arrows point can be inverted. You can adjust the length
-	 * (in pixels) of how far the arrows extend away from the line.
- 	 *
-	 * @param drawDirectionalArrows to enable or disable
-	 * @param invertDirection invert the direction the arrows are drawn. Use null for default value
-	 */
-	void setDrawDirectionalArrows(
-			boolean drawDirectionalArrows, Boolean invertDirection, float strokeWidth) {
-		this.mDrawDirectionalArrows = drawDirectionalArrows;
-		// reset defaults if disabling
-		if (!drawDirectionalArrows) {
-			mInvertDirectionalArrows = DEFAULT_INVERT_ARROWS;
-			return;
-		}
-		setStrokeWidth(strokeWidth);
-		if (invertDirection != null) {
-			this.mInvertDirectionalArrows = invertDirection;
-		}
-	}
-
-	ArrayList<Path> getDirectionalArrowPaths() {
-		return mDirectionalArrows;
 	}
 
 	/**
@@ -155,9 +129,6 @@ class LinearRing implements SegmentClipper.SegmentClippable{
 
 		final List<RectL> segments = new ArrayList<>();
 
-		if (!mDirectionalArrows.isEmpty()) {
-			mDirectionalArrows.clear();
-		}
 		getSegmentsFromProjected(pProjection, mProjectedPoints, segments, pClosePath);
 		final PointL offset;
 		if (pOffset != null) {
@@ -293,10 +264,6 @@ class LinearRing implements SegmentClipper.SegmentClippable{
 			} else {
 				setCloserPoint(screenPoint0, screenPoint1, worldSize);
 				pSegments.add(new RectL(screenPoint0.x, screenPoint0.y, screenPoint1.x, screenPoint1.y));
-
-				if (mDrawDirectionalArrows) {
-					addDirectionalArrow(screenPoint0, screenPoint1);
-				}
 			}
 
 			// update starting point to next position
@@ -386,7 +353,7 @@ class LinearRing implements SegmentClipper.SegmentClippable{
 	 * we can use the same SegmentClipper instead of constructing a new one at each canvas draw.
 	 */
 	public void setClipArea(final long pXMin, final long pYMin, final long pXMax, final long pYMax) {
-		mSegmentClipper = new SegmentClipper(pXMin, pYMin, pXMax, pYMax, this);
+		mSegmentClipper = new SegmentClipper(pXMin, pYMin, pXMax, pYMax, mSegmentClipperChild);
 	}
 
 	/**
@@ -409,109 +376,5 @@ class LinearRing implements SegmentClipper.SegmentClippable{
 		// TODO: Not sure if this is the correct approach
 		this.isHoritonalRepeating = pMapView.isHorizontalMapRepetitionEnabled();
 		this.isVerticalRepeating = pMapView.isVerticalMapRepetitionEnabled();
-	}
-
-	private void addDirectionalArrow(PointL screenPoint0, PointL screenPoint1) {
-		// if the points are really close don't draw an arrow
-		if (Math.abs(screenPoint0.x - screenPoint1.x) +
-				Math.abs(screenPoint0.y - screenPoint1.y) <= mDirectionalArrowLength) {
-			return;
-		}
-
-		if (mInvertDirectionalArrows) {
-			PointL temp = screenPoint0;
-			screenPoint0 = screenPoint1;
-			screenPoint1 = temp;
-		}
-
-		// mid point in projected pixels
-		PointL screenPoint3 = new PointL();
-		PointL screenPoint4 = new PointL();
-		screenPoint3.x = (screenPoint0.x + screenPoint1.x) / 2;
-		screenPoint3.y = (screenPoint0.y + screenPoint1.y) / 2;
-		double delta_x = screenPoint1.x - screenPoint0.x;
-		double delta_y = screenPoint1.y - screenPoint0.y;
-
-		float distance = mDirectionalArrowLength;
-
-		if (delta_x == 0) {
-			screenPoint3.x = screenPoint1.x;
-			screenPoint4.x = screenPoint1.x;
-
-			if (screenPoint0.y > screenPoint1.y) {
-				screenPoint3.y = screenPoint0.y + (int) Math.round(delta_y/2);
-				screenPoint4.y = screenPoint3.y - (int) distance;
-			} else {
-				screenPoint3.y = screenPoint0.y + (int) Math.round(delta_y/2);
-				screenPoint4.y = screenPoint3.y + (int) distance;
-			}
-		} else if (delta_y == 0) {
-			screenPoint3.y = screenPoint1.y;
-			screenPoint4.y = screenPoint1.y;
-
-			if (screenPoint0.x > screenPoint1.x) {
-				screenPoint3.x = screenPoint0.x + (int) Math.round(delta_x/2);
-				screenPoint4.x = screenPoint3.x - (int) distance;
-			} else {
-				screenPoint3.x = screenPoint0.x + (int) Math.round(delta_x/2);
-				screenPoint4.x = screenPoint3.x + (int) distance;
-			}
-		} else {
-			 double slope = delta_y / delta_x;
-			 /*  The formula will calculate a new X,Y coordinate that is some distance away from
-				a given coordinate when a slope is known. Distance is the correct way of
-				thinking about what the number is in regards to the formula used. It can also
-				be thought of as the length of the squares diagonal.
-			*/
-			double r = Math.sqrt(1 + Math.pow(slope, 2));
-			// move pt3 half the distance of the square so pt 5 and 6 will intersect the midpt
-			screenPoint3.x = screenPoint3.x + (int) Math.round((distance / 2) / r);
-			screenPoint3.y = screenPoint3.y + (int) Math.round((distance / 2) * slope / r);
-			//calculate another point on the line that is distance in px away from the mid point
-			screenPoint4.x = screenPoint3.x - (int) Math.round((distance / r));
-			screenPoint4.y = screenPoint3.y - (int) Math.round((distance * slope / r));
-		}
-		/* 3rd point in the square
-			= ( ( x1 + x3 + y1 - y3 ) / 2 , ( x3 - x1 + y1 + y3 ) / 2 )
-		*/
-		PointL screenPoint5 = new PointL();
-		screenPoint5.x =
-				(screenPoint4.x + screenPoint3.x + screenPoint4.y - screenPoint3.y) / 2;
-		screenPoint5.y =
-				(screenPoint3.x - screenPoint4.x + screenPoint4.y + screenPoint3.y) / 2;
-		/* 4th point in the square
-			= ( ( x1 + x3 + y3 - y1 ) / 2 , ( x1 - x3 + y1 + y3 ) / 2 )
-		*/
-		PointL screenPoint6 = new PointL();
-		screenPoint6.x =
-				(screenPoint4.x + screenPoint3.x + screenPoint3.y - screenPoint4.y) / 2;
-		screenPoint6.y =
-				(screenPoint4.x - screenPoint3.x + screenPoint3.y + screenPoint4.y) / 2;
-
-		Path directionalArrowPath = new Path();
-		if (screenPoint0.x > screenPoint1.x && screenPoint0.y > screenPoint1.y) {
-			directionalArrowPath.moveTo(screenPoint5.x, screenPoint5.y);
-			directionalArrowPath.lineTo(screenPoint3.x, screenPoint3.y);
-			directionalArrowPath.lineTo(screenPoint6.x, screenPoint6.y);
-		} else if (screenPoint0.x < screenPoint1.x && screenPoint0.y < screenPoint1.y) {
-			directionalArrowPath.moveTo(screenPoint5.x, screenPoint5.y);
-			directionalArrowPath.lineTo(screenPoint4.x, screenPoint4.y);
-			directionalArrowPath.lineTo(screenPoint6.x, screenPoint6.y);
-		} else if (screenPoint0.x < screenPoint1.x && screenPoint0.y > screenPoint1.y) {
-			directionalArrowPath.moveTo(screenPoint5.x, screenPoint5.y);
-			directionalArrowPath.lineTo(screenPoint4.x, screenPoint4.y);
-			directionalArrowPath.lineTo(screenPoint6.x, screenPoint6.y);
-		} else {
-			directionalArrowPath.moveTo(screenPoint5.x, screenPoint5.y);
-			directionalArrowPath.lineTo(screenPoint3.x, screenPoint3.y);
-			directionalArrowPath.lineTo(screenPoint6.x, screenPoint6.y);
-		}
-		directionalArrowPath.close();
-
-		mDirectionalArrows.add(directionalArrowPath);
-	}
-
-	public void setStrokeWidth(float strokeWidth) {
-		this.mDirectionalArrowLength = DEFAULT_ARROW_LENGTH + strokeWidth;
 	}
 }
